moved code from config/cloud_strategies to cloud_strategy
